Change the package name.
- Changed the package name to add smsotp section to the path.
- Use the orbit version of the packed JARs.
