+++ components/org.wso2.carbon.identity.local.auth.smsotp.authenticator/pom.xml
@@ -135,8 +135,10 @@
                             org.apache.commons.logging; version="${org.apache.commons.logging.range}",
                             org.osgi.framework; version="${osgi.framework.imp.pkg.version.range}",
                             org.osgi.service.component; version="${osgi.service.component.imp.pkg.version.range}",
+
                             org.wso2.carbon.core.*; version="${carbon.kernel.package.import.version.range}",
                             org.wso2.carbon.extension.identity.helper.*; version="${identity.extension.utils.import.version.range}",
+
                             org.wso2.carbon.identity.application.authentication.framework.*; version="${carbon.identity.framework.imp.pkg.version.range}",
                             org.wso2.carbon.identity.application.mgt.*; version="${carbon.identity.framework.imp.pkg.version.range}",
                             org.wso2.carbon.identity.application.common.*; version="${carbon.identity.framework.imp.pkg.version.range}",
@@ -150,16 +152,24 @@
                             org.wso2.carbon.identity.event.event; version="${carbon.identity.framework.imp.pkg.version.range}",
                             org.wso2.carbon.identity.event.services; version="${carbon.identity.framework.imp.pkg.version.range}",
                             org.wso2.carbon.identity.event.handler.notification; version="${identity.event.handler.notification.imp.pkg.version.range}",
-                            org.wso2.carbon.identity.handler.event.account.lock.*; version="${carbon.identity.account.lock.handler.imp.pkg.version.range}",
                             org.wso2.carbon.identity.governance.*; version="${identity.governance.imp.pkg.version.range}",
+                            org.wso2.carbon.identity.handler.event.account.lock.*; version="${carbon.identity.account.lock.handler.imp.pkg.version.range}",
                             org.wso2.carbon.identity.notification.sender.tenant.config.dto; version="${identity.event.handler.notification.imp.pkg.version.range}",
                             org.wso2.carbon.identity.notification.sender.tenant.config; version="${identity.event.handler.notification.imp.pkg.version.range}",
+
                             org.wso2.carbon.idp.mgt.*; version="${carbon.identity.framework.imp.pkg.version.range}",
+
                             org.wso2.carbon.user.api.*; version="${carbon.user.api.imp.pkg.version.range}",
                             org.wso2.carbon.user.core.*; version="${carbon.kernel.package.import.version.range}",
+
                             org.wso2.carbon.utils.*; version="${carbon.kernel.package.import.version.range}",
+
                             org.owasp.encoder; version="${encoder.wso2.import.version.range}",
                             javax.servlet.http; version="${imp.pkg.version.javax.servlet}",
+
+                            org.wso2.carbon.identity.auth.otp.core; version="${identity.auth.otp.commons.version.range}",
+                            org.wso2.carbon.identity.auth.otp.core.constant; version="${identity.auth.otp.commons.version.range}",
+                            org.wso2.carbon.identity.auth.otp.core.model; version="${identity.auth.otp.commons.version.range}",
                         </Import-Package>
                     </instructions>
                 </configuration>
