+++ components/org.wso2.carbon.identity.local.auth.smsotp.authenticator/src/main/java/org/wso2/carbon/identity/local/auth/smsotp/authenticator/exception/SMSOTPAuthenticatorServerException.java
@@ -16,16 +16,14 @@
  * under the License.
  */
 
-package org.wso2.carbon.identity.local.auth.authenticator.exception;
-
-import org.wso2.carbon.identity.base.IdentityException;
+package org.wso2.carbon.identity.local.auth.smsotp.authenticator.exception;
 
 /**
- * SMS OTP authenticator exception implementation.
+ * Implementation of server errors.
  */
-public class SMSOTPAuthenticatorException extends IdentityException {
+public class SMSOTPAuthenticatorServerException extends SMSOTPAuthenticatorException {
 
-    private static final long serialVersionUID = -6133941275898271108L;
+    private static final long serialVersionUID = -4635116841173579773L;
 
     /**
      * Constructs a new exception with an error code, detail message and throwable.
@@ -34,7 +32,7 @@ public class SMSOTPAuthenticatorException extends IdentityException {
      * @param message   The detail message.
      * @param throwable Throwable.
      */
-    public SMSOTPAuthenticatorException(String errorCode, String message, Throwable throwable) {
+    public SMSOTPAuthenticatorServerException(String errorCode, String message, Throwable throwable) {
 
         super(errorCode, message, throwable);
         this.setErrorCode(errorCode);
