+++ pom.xml
@@ -64,8 +64,8 @@
 
         <encoder.wso2.version>1.2.0.wso2v1</encoder.wso2.version>
 
-        <com.vonage.client.version>7.9.0</com.vonage.client.version>
-        <com.twilio.sdk.version>9.13.1</com.twilio.sdk.version>
+        <com.vonage.client.version>7.10.0.wso2v1</com.vonage.client.version>
+        <com.twilio.sdk.version>9.14.0.wso2v1</com.twilio.sdk.version>
 
         <org.apache.commons.logging.range>[1.2.0,2.0.0)</org.apache.commons.logging.range>
         <osgi.framework.imp.pkg.version.range>[1.7.0, 2.0.0)</osgi.framework.imp.pkg.version.range>
@@ -81,6 +81,7 @@
         </carbon.identity.framework.imp.pkg.version.range>
         <identity.event.handler.notification.imp.pkg.version.range>[1.3.15, 2.0.0)
         </identity.event.handler.notification.imp.pkg.version.range>
+        <identity.auth.otp.commons.version.range>[1.0.0, 2.0.0)</identity.auth.otp.commons.version.range>
     </properties>
 
     <dependencyManagement>
@@ -200,13 +201,13 @@
             </dependency>
 
             <dependency>
-                <groupId>com.twilio.sdk</groupId>
-                <artifactId>twilio</artifactId>
+                <groupId>org.wso2.orbit.com.twilio.sdk</groupId>
+                <artifactId>twilio.wso2</artifactId>
                 <version>${com.twilio.sdk.version}</version>
             </dependency>
             <dependency>
-                <groupId>com.vonage</groupId>
-                <artifactId>client</artifactId>
+                <groupId>org.wso2.orbit.com.vonage</groupId>
+                <artifactId>client.wso2</artifactId>
                 <version>${com.vonage.client.version}</version>
             </dependency>
